Clean up code, remove icons
