+++ AuroraEditor/Base/CommandPalette/Model/Command.swift
@@ -10,17 +10,15 @@ import Foundation
 
 class Command: ObservableObject, Identifiable, Hashable {
     static func == (lhs: Command, rhs: Command) -> Bool {
-        lhs.name == rhs.name && lhs.icon == rhs.icon
+        lhs.id == rhs.id
     }
 
     @Published var name: String
-    @Published var icon: String
     var command: () -> Void = {}
     var id = UUID()
 
-    init(name: String, icon: String, command: @escaping () -> Void = {}) {
+    init(name: String, command: @escaping () -> Void = {}) {
         self.name = name
-        self.icon = icon
         self.command = command
     }
