+++ AuroraEditor/Base/CommandPalette/UI/CommandPaletteItem.swift
@@ -19,16 +19,10 @@ struct CommandPaletteItem: View {
 
     public var body: some View {
         HStack(spacing: 8) {
-            Image(systemName: command.icon)
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 20, height: 20)
-                .padding(2)
-                .padding(.leading, 5)
             VStack(alignment: .leading, spacing: 0) {
                 Text(command.name).font(.system(size: 14))
                     .lineLimit(1)
-            }.padding(.trailing, 15)
+            }.padding(.horizontal, 15)
             Spacer()
         }
     }
